edit

+++ src/components/Auth/LoginPage.js
@@ -49,7 +49,10 @@ const LoginPage = () => {
   };
 
   useEffect(() => {
-    localStorage.clear();
+    let token = localStorage.getItem("token");
+    if (token) {
+      localStorage.clear();
+    }
   }, []);
 
   return (
